need better tests but impl first
